Simplify updating a subject
This commit includes structural changes in the way the subject is
updated. It uses destructuring and universal updateSubject function
for updating both name and scores. The actions are built with resepct to
Flux Standard Action, but don't contain meta property.
This is due to this topic:
https://github.com/acdlite/flux-standard-action/issues/18

id and subjectsType are essential for all the actions so they're kept
in payload